+++ pysep/__init__.py
@@ -1,6 +1,8 @@
 import logging
+from importlib.metadata import version
 
-__version__ = "0.4.1"
+# Defines version number from 'pyproject.toml'
+__version__ = version("pysep-adjtomo")
 
 logger = logging.getLogger("pysep")
 logger.setLevel("INFO")
@@ -15,3 +17,4 @@ logger.addHandler(ch)
 from pysep.pysep import Pysep, get_data  # NOQA
 from pysep.recsec import RecordSection, plotw_rs  # NOQA
 from pysep.declust import Declust  # NOQA
+from pysep.utils.io import read_sem, read_stations, read_events_plus  # NOQA